requirements.txt
# 외국인등록증 뒷면 분류기 - 의존성 패키지

# 핵심 머신러닝 라이브러리
tensorflow>=2.13.0,<2.16.0
numpy>=1.21.0,<1.25.0
opencv-python>=4.7.0,<4.9.0
pillow>=9.0.0,<11.0.0

# 데이터 처리 및 시각화
matplotlib>=3.6.0,<3.8.0
seaborn>=0.11.0,<0.13.0
scikit-learn>=1.2.0,<1.4.0
pandas>=1.5.0,<2.1.0

# 개발 도구
jupyter>=1.0.0
tqdm>=4.64.0

# TensorFlow.js 변환
tensorflowjs>=4.0.0,<5.0.0

# 웹 서버 (선택사항)
flask>=2.3.0,<3.0.0
flask-cors>=4.0.0

# 호환성 패키지 (Windows 환경)
pywin32>=300; sys_platform == "win32"

# 추가 유틸리티
requests>=2.28.0
